requirements_unix.txt
aiofiles==23.2.1
aiogram==3.4.1
aiohttp==3.9.3
aiosignal==1.3.1
alembic==1.13.1
annotated-types==0.6.0
async-timeout==4.0.3
asyncpg==0.29.0
attrs==23.2.0
cachetools==5.3.3
certifi==2024.2.2
environ==1.0
environs==11.0.0
frozenlist==1.4.1
greenlet==3.0.3
idna==3.6
magic-filter==1.0.12
Mako==1.3.3
MarkupSafe==2.1.5
marshmallow==3.21.1
multidict==6.0.5
packaging==24.0
psycopg2-binary
pydantic==2.5.3
pydantic_core==2.14.6
python-dotenv==1.0.1
redis==5.0.3
SQLAlchemy==2.0.29
typing_extensions==4.10.0
yarl==1.9.4
